@@ -17,6 +17,7 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using Catalog.API.Extensions;
 using Catalog.API.Infrastructure.AutofacModules;
+using Microsoft.Extensions.Hosting;
 
 namespace Catalog.API
 {
@@ -32,12 +33,33 @@ namespace Catalog.API
         // This method gets called by the runtime. Use this method to add services to the container.
         public IServiceProvider ConfigureServices(IServiceCollection services)
         {
+            
            services.AddCustomMVC(Configuration)
                 .AddCustomMongoDBContext()
                 .AddCustomOptions(Configuration)
                 .AddIntegrationServices(Configuration)
                 .AddEventBus(Configuration)
                 .AddSwagger();
+
+            services.AddCap(x => {
+                //x.UseDashboard();
+                x.UseMongoDB(o => {
+                    o.DatabaseConnection = "mongodb://admin:12345678@127.0.0.1:27017,127.0.0.1:27018,127.0.0.1:27019/?connectTimeoutMS=10000&authSource=admin&authMechanism=SCRAM-SHA-1";
+                    o.DatabaseName = "SampleCatalog";
+                    o.PublishedCollection = "cap.published";
+                    o.ReceivedCollection = "cap.received";
+                });
+                x.DefaultGroup = "sampleCatalog";
+                x.UseRabbitMQ(o => {
+                    o.HostName = "127.0.0.1";
+                    o.Port = 5672;
+                    o.UserName = "admin";
+                    o.Password = "admin123456";
+                    o.VirtualHost = "frameworksample";
+                    o.ExchangeName = "frameworksample-exchange";
+                });
+
+            });
 
             var container = new ContainerBuilder();
             container.Populate(services);
@@ -47,7 +69,7 @@ namespace Catalog.API
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IHostEnvironment env)
         {
             if (env.IsDevelopment())
             {
@@ -66,10 +88,15 @@ namespace Catalog.API
             }
 
             app.UseCors("CorsPolicy");
-            app.UseMvcWithDefaultRoute();
+            
 
             app.UseHttpsRedirection();
-            app.UseMvc();
+
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints => {
+                endpoints.MapControllers();
+            });
 
             app.UseSwagger()
                 .UseSwaggerUI(c =>
