@@ -26,23 +26,23 @@ namespace Auth.ClientTestApi
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddMvc()
-                //全局配置Json序列化处理
-                .AddJsonOptions(options =>
-                {
-                    //忽略循环引用
-                    options.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
-                    //设置时间格式
-                    //options.SerializerSettings.DateFormatString = "yyyy-MM-dd";
-                    //不使用驼峰样式的key
-                    //options.SerializerSettings.ContractResolver = new DefaultContractResolver();
+                ////全局配置Json序列化处理
+                //.AddJsonOptions(options =>
+                //{
+                //    //忽略循环引用
+                //    options.SerializerSettings.ReferenceLoopHandling = ReferenceLoopHandling.Ignore;
+                //    //设置时间格式
+                //    //options.SerializerSettings.DateFormatString = "yyyy-MM-dd";
+                //    //不使用驼峰样式的key
+                //    //options.SerializerSettings.ContractResolver = new DefaultContractResolver();
 
-                    //自定义序列化
-                    var convers = options.SerializerSettings.Converters ?? new List<JsonConverter>();
-                    convers.Add(new StringEnumConverter());
+                //    //自定义序列化
+                //    var convers = options.SerializerSettings.Converters ?? new List<JsonConverter>();
+                //    convers.Add(new StringEnumConverter());
 
-                    options.SerializerSettings.Converters = convers;
-                })
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+                //    options.SerializerSettings.Converters = convers;
+                //})
+                .SetCompatibilityVersion(CompatibilityVersion.Version_3_0);
 
             services.AddCustomAuthentication(Configuration);
             services.AddCustomSwagger(Configuration);
